fix after review: renaming, minor code changes
