+++ process/throttle/antiflood/disabled/nilTopicFloodPreventer_test.go
@@ -1,4 +1,4 @@
-package floodPreventers
+package disabled
 
 import (
 	"fmt"
